+++ src/main/java/com/jug/lp/LpUtils.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import net.imglib2.algorithm.componenttree.Component;
-import net.imglib2.type.numeric.real.DoubleType;
+import net.imglib2.type.numeric.real.FloatType;
 
 import com.jug.util.ComponentTreeUtils;
 
@@ -20,7 +20,7 @@ public class LpUtils {
 	 * Builds and returns the set Hup (given as a List). Hup is defines as the
 	 * set up all Hypothesis in hyps that strictly above the Hypothesis hyp (in
 	 * image space).
-	 *
+	 * 
 	 * @param hyp
 	 *            the reference hypothesis.
 	 * @param hyps
@@ -28,9 +28,9 @@ public class LpUtils {
 	 * @return a List of all Hypothesis from hyps that lie above the
 	 *         segmentation hypothesis hyp.
 	 */
-	public static List< Hypothesis< Component< DoubleType, ? >>> getHup( final Hypothesis< Component< DoubleType, ? >> hyp, final List< Hypothesis< Component< DoubleType, ? >>> hyps ) {
-		final List< Hypothesis< Component< DoubleType, ? >>> Hup = new ArrayList< Hypothesis< Component< DoubleType, ? >>>();
-		for ( final Hypothesis< Component< DoubleType, ? >> candidate : hyps ) {
+	public static List< Hypothesis< Component< FloatType, ? >>> getHup( final Hypothesis< Component< FloatType, ? >> hyp, final List< Hypothesis< Component< FloatType, ? >>> hyps ) {
+		final List< Hypothesis< Component< FloatType, ? >>> Hup = new ArrayList< Hypothesis< Component< FloatType, ? >>>();
+		for ( final Hypothesis< Component< FloatType, ? >> candidate : hyps ) {
 			if ( ComponentTreeUtils.isAbove( candidate.getWrappedHypothesis(), hyp.getWrappedHypothesis() ) ) {
 				Hup.add( candidate );
 			}
